Split BinderHub and JupyterHub config files
