+++ testing/local-binder-local-hub/api_token.txt
@@ -1,0 +1,1 @@
+secret123
